Figured out the issue with problem.md in trouble
Basically, a macro match was gobbling it up. Refactoring that part would
be substantial and is a goal, but it will take time. In the meantime, if
you have something like CAPS(_"dude") then change that to CAPS
(_"dude"). This is true even if CAPS does not match anything. That's the
problem.

Did find out that the parsing is not line based in compiling code
blocks.